+++ config/src/main/java/org/jboss/hal/config/keycloak/Keycloak.java
@@ -15,106 +15,93 @@
  */
 package org.jboss.hal.config.keycloak;
 
-import jsinterop.annotations.JsFunction;
+import elemental2.promise.Promise;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 import static jsinterop.annotations.JsPackage.GLOBAL;
+import static org.jboss.hal.config.keycloak.KeycloakSingleton.TOKEN_TIMEOUT;
+import static org.jboss.hal.resources.UIConstants.OBJECT;
 
 /**
  * https://www.keycloak.org/docs/latest/securing_apps/index.html#_javascript_adapter
  */
-@JsType(namespace = GLOBAL, isNative = true)
+@JsType(isNative = true, namespace = GLOBAL)
 public class Keycloak {
 
     public String subject;
     public String token;
+    public KeycloakRoles realmAccess;
+    public KeycloakProfile profile;
 
-    public RealmAccess realmAccess;
-
-    public UserProfile userProfile;
-
-    public Keycloak(String kcConfigUrl) {
+    public Keycloak(KeycloakConfig config) {
     }
 
-    public native Api init(Api options);
-
     /**
-     * Redirects to logout.
-     *
-     * @param options Options is an Object, where: redirectUri - Specifies the uri to redirect to after logout.
-     *
+     * Initializes this Keycloak instance by calling the {@link #init(KeycloakInitOptions)} method, then the
+     * {@link #updateToken(int)} method and finally the {@link #loadUserProfile()} method. The user profile is assigned to the
+     * {@link #profile} field.
      */
-    public native String logout(String options);
-
-    /**
-     * Returns the URL to logout the user.
-     *
-     * @param options Options is an Object, where: redirectUri - Specifies the uri to redirect to after logout.
-     *
-     */
-    public native String createLogoutUrl(String options);
-
-    /** Returns the URL of the Account Management Console in keycloak server. */
-    public native String createAccountUrl();
-
-    public native UserProfile loadUserProfile();
-
-    public native void updateToken(int time);
-
-    @JsType(name = "options")
-    public static class Api {
-
-        public String onLoad = "login-required";
+    @JsOverlay
+    public final Promise<Boolean> authenticate() {
         // the default responseMode is "fragment" but there is a problem as the keycloak javascript appends the
         // token in the url fragment and as it is parsed in the ParameterTokenFormatter, it uses the '=' char to split
         // each gwt place parameter and doesn't consider the keycloak values, then throws an exception because there is
         // key=value pair
-        public String responseMode = "query";
+        KeycloakInitOptions options = new KeycloakInitOptions();
+        options.onLoad = "login-required";
+        options.responseMode = "query";
 
-        public native Api success(SuccessCallback authenticated);
-
-        public native Api error(ErrorCallback callback);
-
-        @JsFunction
-        public interface SuccessCallback {
-            void success(boolean authenticated);
-        }
-
-        @JsFunction
-        public interface ErrorCallback {
-            void error();
-        }
+        return init(options)
+                .then(authenticated -> updateToken(TOKEN_TIMEOUT))
+                .then(ignore -> loadUserProfile())
+                .then(profile -> {
+                    this.profile = profile;
+                    return Promise.resolve(true);
+                });
     }
 
-    @JsType(isNative = true, namespace = GLOBAL)
-    public static class UserProfile {
+    public native Promise<Boolean> init(KeycloakInitOptions options);
 
+    /** Returns the URL of the Account Management Console in keycloak server. */
+    public native String createAccountUrl();
+
+    public native Promise<KeycloakProfile> loadUserProfile();
+
+    public native Promise<Boolean> updateToken(int time);
+
+    public native String logout(Object options);
+
+    // ------------------------------------------------------ inner classes
+
+    @JsType(isNative = true, namespace = GLOBAL, name = OBJECT)
+    public static class KeycloakConfig {
+
+        public String url;
+        public String realm;
+        public String clientId;
+    }
+
+    @JsType(isNative = true, namespace = GLOBAL, name = OBJECT)
+    public static class KeycloakInitOptions {
+
+        String onLoad;
+        String responseMode;
+    }
+
+    @JsType(isNative = true, namespace = GLOBAL, name = OBJECT)
+    public static class KeycloakRoles {
+
+        public String[] roles;
+    }
+
+    @JsType(isNative = true, namespace = GLOBAL, name = OBJECT)
+    public static class KeycloakProfile {
+
+        public String id;
+        public String username;
         public String email;
         public String firstName;
         public String lastName;
-        public String username;
-
-        public native void success(SuccessCallback profile);
-
-        @JsOverlay
-        @Override
-        public final String toString() {
-            return "UserProfile{" + firstName + " " + lastName + ", " + "email='" + email + '\'' + ", username='" +
-                    username + '}';
-        }
-
-        @JsFunction
-        public interface SuccessCallback {
-            void success(UserProfile profile);
-        }
     }
-
-    @JsType(isNative = true, namespace = GLOBAL)
-    public static class RealmAccess {
-
-        public String[] roles;
-
-    }
-
 }